requirements.txt
alembic==1.11.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
toml==0.10.2
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
alembic==1.11.1
azure-core==1.29.1
azure-storage-blob==12.17.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
toml==0.10.2
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
alembic==1.11.1
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.29.1
azure-identity==1.14.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.17.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.0.5
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
pyparsing==3.1.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
toml==0.10.2
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
